feat: update and move popup, update localhost to persistentAtom, use client:only for preact and add loading spinner
