+++ src/pages/biomon.astro
@@ -9,7 +9,13 @@ import '../styles/biomon.css';
 <HTML title="IMI GLOBAL — Biomonitor" bodyClass="charsheet-page biomon-page">
   <Layout>
     <Charsheet>
-      <Biomonitor client:load />
+      <Biomonitor client:only="preact">
+        <div slot="fallback" class="loading-fallback">
+          <svg class="loading-spinner" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
+            <circle cx="12" cy="12" r="10" fill="none" stroke="currentColor" stroke-width="2" stroke-dasharray="31.4 31.4" stroke-linecap="round" />
+          </svg>
+        </div>
+      </Biomonitor>
     </Charsheet>
   </Layout>
 </HTML>
