+++ src/styles/global/utilities.css
@@ -157,3 +157,22 @@
   border-style: solid;
   border-color: var(--accent-muted);
 }
+
+/* === Loading Fallback === */
+.loading-fallback {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  min-height: 120px;
+}
+
+.loading-spinner {
+  width: 24px;
+  height: 24px;
+  color: var(--accent-muted);
+  animation: spin 1s linear infinite;
+}
+
+@keyframes spin {
+  to { transform: rotate(360deg); }
+}
